+++ roguelike/melee_weapon.rb
@@ -11,14 +11,16 @@ class MeleeWeapon
       bonus_health: 50,
       bonus_mana: 50,
       bonus_strength: 10,
-      bonus_self_regen: 10
+      bonus_self_regen: 10,
+      description: "A powerful sword said to have been lost in time."
     })
     new({
       name: 'Rusty Dagger',
       icon: '†',
       equipment_slot: :main_hand,
       weight: 4,
-      bonus_strength: 1
+      bonus_strength: 1,
+      description: "A standard dagger that appears to be quite aged."
     })
     new({
       name: 'Fire Sword',
@@ -27,7 +29,8 @@ class MeleeWeapon
       equipment_slot: :main_hand,
       weight: 4,
       bonus_strength: 1,
-      on_hit_effect: Evals.new_dot(5, 2, 'fire', false)
+      on_hit_effect: Evals.new_dot(5, 2, 'fire', false),
+      description: "A sword imbued with flame. Will burn enemies for a short time, dealing damage per tick."
     })
   end
 end