+++ roguelike/spell.rb
@@ -13,9 +13,9 @@ class Spell
     end
     if Player.mana >= mana_cost
       if is_projectile
-        Settings.ready('cast', self, self.range)
+        Settings.ready('cast', self.dup, self.range)
       else
-        Settings.ready_cast(self, self.range || 0)
+        Settings.ready_cast(self.dup, self.range || 0)
       end
     else
       Log.add "Not enough Mana."
@@ -37,7 +37,8 @@ class Spell
       projectile_speed: 20,
       on_hit_damage: 0,
       collided_action: Evals.explode(1, 20, 'fire'),
-      mana_cost: 5
+      mana_cost: 5,
+      description: "A slow moving projectile spell that will explode with a radius of 1 on impact or upon reaching it's furthest point. Deals immediate fire damage to all within the explosion."
     })
     new({
       name: 'Fire Ball',
@@ -50,7 +51,8 @@ class Spell
       projectile_speed: 20,
       on_hit_damage: 0,
       collided_action: Evals.explode(0, 10, 'fire'),
-      mana_cost: 3
+      mana_cost: 3,
+      description: "A slow moving projectile spell that will explode upon impact or upon reaching it's furthest point. Deals immediate fire damage to the collided entity."
     })
     new({
       name: 'Poison Blast',
@@ -63,7 +65,8 @@ class Spell
       projectile_speed: 20,
       on_hit_damage: 0,
       collided_action: Evals.new_dot(5, 2, 'poison'),
-      mana_cost: 2
+      mana_cost: 2,
+      description: "A slow moving projectile spell that will poison the collided entity, dealing poison damage over time."
     })
     new({
       name: 'Berserk',
@@ -72,7 +75,8 @@ class Spell
       is_projectile: false,
       non_projectile_script: Evals.player_berserk(50),
       type: 'physical',
-      mana_cost: 5
+      mana_cost: 5,
+      description: "Grants berserk for 50 ticks. While Berserk, the Player will have +3 speed and +50% strength. Berserk will cost 1 energy per tick and will cancel early if the Player runs out of energy."
     })
     new({
       name: 'Summon Stone',
@@ -82,7 +86,8 @@ class Spell
       range: 5,
       non_projectile_script: Evals.summon_stone,
       type: 'physical',
-      mana_cost: 5
+      mana_cost: 5,
+      description: "Will summon a solid block at target location within range."
     })
   end
 end