+++ roguelike/changelog.md
@@ -1,4 +1,12 @@
-Oct 14, 205 v0.0.59
+Oct 15, 2015 v0.0.60
+Reduced gold value of drops and findings
+Changed the way the slimes decide to split
+Slimes now drop the amount of slime balls equal to their strength/size
+Fixed game-crashing bug caused when trying to look at a slime in any targeting menu.
+Player no longer picks up all of the items he dropped if he performs an activity in the same spot
+
+
+Oct 14, 2015 v0.0.59
 Added type, mana, and range for spells when casting them
 Can now 'Read More' on Spell in the SpellBook/Amulets
 'Read More' only shows stats when it's equipment and shows spells if it can cast spells
@@ -22,7 +30,7 @@ Non-projectile spells can now be cast
   Spells with range will automatically open in the Targeting menu
 Added a range indicator to the targeting menu
 
-Oct 12, 2015 v.0.0.57
+Oct 12, 2015 v0.0.57
 Bonuses are made to be faster in code
 Added a Maze level every 5 levels